[추가] 테스트 케이스 추가 및 수정
- 나뉘어진 Controller와 추가된 로직들에 대해서 테스트 케이스를 추가하였다.
- 추가한 테스트 케이스에 따라 수정이 필요한 부분들을 수정하였다.
