+++ src/main/java/webserver/controllers/ResolveController.java
@@ -11,19 +11,17 @@ import java.util.Map;
 public class ResolveController {
     private final static ResolveController RESOLVE_CONTROLLER = new ResolveController();
     private final static Logger logger = LoggerFactory.getLogger(ResolveController.class);
+    private final static Controller staticFileController = new StaticFileController();
     private final static Map<String, Controller> requestControllers = new HashMap<>() {{
         put("/user/create", new UserCreateController());
-        put("static", new StaticFileController());
+        put("static", staticFileController);
     }};
 
     public static ResolveController getInstance() {
         return RESOLVE_CONTROLLER;
     }
 
-    public Controller resolveRequest(HttpRequest request) throws IOException {
-        if (requestControllers.containsKey(request.path())) {
-            return requestControllers.get(request.path());
-        }
-        return requestControllers.get("static");
+    public Controller resolveRequest(HttpRequest request) {
+        return requestControllers.getOrDefault(request.path(), staticFileController);
     }
 }
